+++ packet.odin
@@ -45,7 +45,7 @@ Initial_Packet :: struct {
 	dest_conn_id:   Connection_Id,
 	source_conn_id: Connection_Id,
 	token:          []u8,
-	packet_number:  u32,
+	packet_number:  u64,
 	packet_payload: []^Frame,
 }
 
@@ -53,7 +53,7 @@ Zero_RTT_Packet :: struct {
 	version:        u32, // version is 0 in case of negotiation
 	dest_conn_id:   Connection_Id,
 	source_conn_id: Connection_Id,
-	packet_number:  u32,
+	packet_number:  u64,
 	packet_payload: []^Frame,
 }
 
@@ -61,7 +61,7 @@ Handshake_Packet :: struct {
 	version:        u32, // version is 0 in case of negotiation
 	dest_conn_id:   Connection_Id,
 	source_conn_id: Connection_Id,
-	packet_number:  u32,
+	packet_number:  u64,
 	packet_payload: []^Frame,
 }
 
@@ -83,7 +83,7 @@ One_RTT_Packet :: struct {
 	spin_bit:       bool,
 	key_phase:      bool,
 	dest_conn_id:   Connection_Id,
-	packet_number:  u32,
+	packet_number:  u64,
 	packet_payload: []^Frame,
 }
